requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
mongoengine==0.28.2
pymongo==4.6.3
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
google-generativeai
